requirements.txt
anyio==3.6.2
asgiref==3.6.0
bcrypt==3.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
motor==2.5.1
passlib==1.7.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.6
PyJWT==2.3.0
pymongo==3.12.1
python-decouple==3.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.16.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.14.0
